Add prompt

+++ src/relock/index.ts
@@ -4,6 +4,7 @@ import shelljs from 'shelljs';
 import findit from 'findit';
 import path from 'path';
 import _ from 'lodash';
+import inquirer from 'inquirer';
 
 shelljs.config.fatal = true;
 
@@ -27,15 +28,30 @@ const findNpmRoots = (): Promise<string[]> => {
   });
 };
 
-findNpmRoots().then((dirs) => {
-  _.forEach(dirs, (dir) => {
-    console.log(`Re-locking ${dir}`);
-    shelljs.cd(dir);
-    console.log('  removing node_modules');
-    shelljs.rm('-rf', 'node_modules');
-    console.log('  removing package-lock.json');
-    shelljs.rm('package-lock.json');
-    console.log('  running npm install');
-    shelljs.exec(`npm i`, { silent: true, fatal: true });
-  });
-});
+const main = async () => {
+  const dirs = await findNpmRoots();
+  console.log('Re-locking following dirs:');
+  dirs.forEach((d) => console.log(`  ${d}`));
+  const { goAhead } = await inquirer.prompt([
+    {
+      type: 'confirm',
+      name: 'goAhead',
+      message: 'Continue?',
+      default: true,
+    },
+  ]);
+  if (goAhead) {
+    _.forEach(dirs, (dir) => {
+      console.log(`Re-locking ${dir}`);
+      shelljs.cd(dir);
+      console.log('  removing node_modules');
+      shelljs.rm('-rf', 'node_modules');
+      console.log('  removing package-lock.json');
+      shelljs.rm('package-lock.json');
+      console.log('  running npm install');
+      shelljs.exec(`npm i`, { silent: true, fatal: true });
+    });
+  }
+};
+
+main().then();
